refactor(Controller, Car, ResultView): refactor mvc pattern
